Mostly working solution without styling
